requirements.txt
accelerate==0.23.0
backcall==0.2.0
bdc-payments==0.1.1
boxsdk==3.9.2
en-core-web-sm==3.7.1
fire==0.5.0
flair==0.12.2
flask-api==3.1
flask-restful==0.3.10
google-api-python-client==2.143.0
html-to-json==2.0.0
ipdb==0.13.13
language-formatters-pre-commit-hooks==2.11.0
loguru==0.7.2
markdown==3.5.1
nlp==0.4.0
nlpaug==1.1.11
pandasai==2.0.18
pickleshare==0.7.5
pip-chill==1.0.3
psycopg2==2.9.9
pybind11==2.11.1
pysocks==1.7.1
replicate==0.15.5
runipy==0.1.5
setfit==1.0.1
snowflake-cli-labs==2.6.1
snowflake-ml-python==1.6.1
snowflake-sqlalchemy==1.5.1
streamlit-ace==0.1.1
textdistance==4.6.1
thefuzz==0.22.1
types-pyyaml==6.0.12.12
types-tabulate==0.9.0.3
tzlocal==5.2
validators==0.22.0
